FEM: Add CalculiX binary usability check
That patch add a simple CalculiX usability check. It runs CalculiX
binary ccx and searches output for string "CalculiX". The failure
to pass is not treated as a critical error, so the user can try to
run calculations even if the check fails. The check is performed
only if python can identify the system as Linux or Windows.

Signed-off-by: Przemo Firszt <przemo@firszt.eu>